added a DateAndTimeService on the frontend to get those functions out of helpers w/ burnettk
