+++ spiffworkflow-frontend/src/routes/MyTasks.tsx
@@ -8,8 +8,6 @@ import {
   getPageInfoFromSearchParams,
   modifyProcessIdentifierForPathParam,
   refreshAtInterval,
-  REFRESH_INTERVAL_SECONDS,
-  REFRESH_TIMEOUT_SECONDS,
 } from '../helpers';
 import HttpService from '../services/HttpService';
 import {
@@ -19,6 +17,7 @@ import {
   RecentProcessModel,
 } from '../interfaces';
 import ProcessInstanceRun from '../components/ProcessInstanceRun';
+import DateAndTimeService from '../services/DateAndTimeService';
 
 const PER_PAGE_FOR_TASKS_ON_HOME_PAGE = 5;
 
@@ -47,8 +46,8 @@ export default function MyTasks() {
 
     getTasks();
     refreshAtInterval(
-      REFRESH_INTERVAL_SECONDS,
-      REFRESH_TIMEOUT_SECONDS,
+      DateAndTimeService.REFRESH_INTERVAL_SECONDS,
+      DateAndTimeService.REFRESH_TIMEOUT_SECONDS,
       getTasks
     );
   }, [searchParams]);
